docs: add DartDoc documentation and optimize file tree widgets
